+++ modules/auxiliary/analyze/hashcat_aix.rb
@@ -78,8 +78,10 @@ class MetasploitModule < Msf::Auxiliary
         end
       end
     end
-    cleanup_files.each do |f|
-      File.delete(f)
+    if datastore['DeleteTempFiles']
+      cleanup_files.each do |f|
+        File.delete(f)
+      end
     end
   end
